Move subproviders into dir, along with random utility modules into utils dir
