+++ lib/blockchain.js
@@ -5,12 +5,13 @@ var Trie = require('merkle-patricia-tree');
 var FakeTransaction = require('ethereumjs-tx/fake.js');
 var utils = require('ethereumjs-util');
 var seedrandom = require('seedrandom');
-var random = require('./random');
 var bip39 = require('bip39')
 //var hdkey = require('ethereumjs-wallet/hdkey')
 var lightwallet = require("eth-lightwallet");
-var Log = require("./log");
-var to = require('./to');
+
+var Log = require("./utils/log");
+var to = require('./utils/to');
+var random = require('./utils/random');
 
 Blockchain = function(options) {
   var self = this;
