+++ lib/manager.js
@@ -1,6 +1,6 @@
 var Blockchain = require('./blockchain.js');
 var Compiler = require('./compiler.js');
-var to = require('./to');
+var to = require('./utils/to');
 var utils = require('ethereumjs-util');
 var fs = require('fs');
 var async = require('async');
