+++ lib/server.js
@@ -5,9 +5,9 @@ var ProviderEngine = require("web3-provider-engine");
 var FilterSubprovider = require('web3-provider-engine/subproviders/filters.js');
 var VmSubprovider = require('web3-provider-engine/subproviders/vm.js');
 
-var RequestFunnel = require('./requestfunnel.js');
-var ReactiveBlockTracker = require("./reactiveblocktracker.js");
-var GethDefaults = require("./gethdefaults.js");
+var RequestFunnel = require('./subproviders/requestfunnel.js');
+var ReactiveBlockTracker = require("./subproviders/reactiveblocktracker.js");
+var GethDefaults = require("./subproviders/gethdefaults.js");
 
 Server = {
   server: function(options) {
