+++ lib/utils/random.js
